+++ mahu-admin/mahu-admin-infra/src/main/java/cool/houge/mahu/admin/security/AuthContext.java
@@ -14,7 +14,7 @@ public interface AuthContext {
     /// 匿名访问
     AuthContext ANONYMOUS = new AuthContext() {
         @Override
-        public long adminId() {
+        public int adminId() {
             throw new UnsupportedOperationException("anonymous");
         }
 
@@ -35,7 +35,7 @@ public interface AuthContext {
     };
 
     /// 返回认证的管理员 ID
-    long adminId();
+    int adminId();
 
     /// 返回认证用户的名称
     String name();
